Improve support for JAX_DUMP_IR_TO.
Previously the environment variable JAX_DUMP_IR_TO controlled
whether and where to dump the MLIR module prior to compilation. Now we move the code for that support from
compiler.py to mlir.py, so that it can be used in other
parts of the code. We also add support for logging to Sponge.

Using this support we now log the module on errors from
refine_polymorphic_shapes.

PiperOrigin-RevId: 592099633