[shape_poly] Cache the hash values for symbolic expressions
Hashing is performance critical for symbolic expressions because
their internal representation is as dictionaries. In some of the
unit tests by caching the hash we save 20% of the calls. Hashing
will be even more important for the upcoming decision procedure
for symbolic expressions.

We also cache the sorting of the monomials in a symbolic expression.