[shape_poly] Minor improvements for handling of dilation in convolutions.
Use core.dilate_dim, it is clearer and especially when using
shape polymoprhism it will result in better error messages.

Replace the use of np.maximum with core.max_dim, because the
latter will result in fewer errors in presence of shape polymorphism
(the core.max_dim is deferring the actual computation to shape
refinement time).