[PJRT:C] Make CreatePjrtApi a regular function instead of constexpr.
This makes CreatePjrtApi more flexible to use. Specifically, I would
like to be able to pass in a reinterpret_cast'd extension pointer in a
future change, and reinterpret_cast cannot be used in constexpr's.

This change also has the callers of CreatePjrtApi to use static
function initialization instead of global contexprs.

PiperOrigin-RevId: 599952707