[shape_poly] Add more tests for reasoning about inequalities.
As I explore more powerful ways to reason about inequalities,
I came up with more tests of inequalities that I wish we can handle.
This PR adds the tests I have so far, even if they do not produce
the correct result yet. I write the expected values for tests as

   _expect(best=v1, current=v2)

to document that the current logic produces `v2` but the best value
we can hope for is `v1`.

This PR also adds more support for profiling tests.