[Mosaic] Use `vector.kind<maximumf>` instead of `vector.kind<maxnumf>`
The semantics are closer to the TPU: having a NaN input results in NaN. However, we don't respect the -0.0 vs +0.0 ordering in older TPUs.

This also fixes a mismatch where we are using `arith.maximumf` for lowering `vector.kind<maxnumf>` (instead of `arith.maximumf` for `vector.kind<maximumf>` or `arith.maxnumf` for `vector.kind<maxnumf>`).

PiperOrigin-RevId: 604849222