[Pallas] Add tpu bitcast op to pallas primitives.
This bitcast is different from the jax.lax.bitcast_convert_type. The former takes takes [m * packing, n] shape while the latter takes [m, n, packing] shape. In Mosaic, we use [m * packing, n] to represent a shape of a vector with narrower type because we always take the last 2 dims of the shape as the tiling dim in Mosaic. In a word, it is hard to directly lower the jax.lax.bitcast_convert_type to tpu.bitcast. Therefore we add it to pallas primitive type.

PiperOrigin-RevId: 605673930