Move test for float0 values out of nary_op_weak_type_rule and into unop_dtype_rule.
With the current code structure, it appears we check for float0s twice for non-unary ops, both in the dtype rule and the weak type rule. The only reason I can see this code is in nary weak type rule is to make sure it is also called for unary operators. But we can achieve that by just putting the code in the unary dtype rule.

It's also a bit odd that this dtype check is in the weak type rule at all: it's just a standard dtype check, so it should go in the dtype rule. This allows us to use the standard weak type rule.

PiperOrigin-RevId: 606671661