[Mosaic] In apply_vector_layout.cc, check layout validity when reading the attribute
This allows us to rely on this throughout the code and replace some checks with TPU_ASSERT_*. They have the semantics of an assert and make it clearer that it is an unexpected internal error (instead of unimplemented or invalid user input that we should handle).

Note: the original error messages for some of these checks were using the wrong input names.
PiperOrigin-RevId: 607463728