[XLA:Python] Port several more modules to nanobind.
Ports:
* xla_compiler
* ops
* custom_call_sharding
* mlir
together with a few miscellaneous changes.

These modules mostly pertain to XLA's computation and sharding types, and form a reasonably coherent change that can be landed by itself. A few pybind11/nanobind translations had to be added, but not too many. The intent is that most of these should go away as part of the next change, which should port the remaining types.

PiperOrigin-RevId: 610593276