Use `$(RULEDIR)` to avoid an implicit dependency on `output_to_genfiles`.
This uses `gentbl_rule` which has `output_to_genfiles = True` set. When removing that, these rules break: genrules by default place their outputs in `bazel-genfiles`, but the input files are now in `bazel-bin`, and the genrule doesn't put it in the right place. Instead of taking the full path, use `$(RULEDIR)` to put in the right place, regardless of where the input file comes from.