make print_saved_residuals show more checkpoint_name names
Previously, we would only show the `checkpoint_name` name for a saved residual
if the _output_ of `checkpoint_name` was saved. That meant that a policy like
`save_dots` could allow a value to be saved, but because the saved version was
the _input_ to a `checkpoint_name`, we wouldn't print the name.

This fix is kind of a quick hack, and doesn't handle all cases. For example, if
a value is an input to two different `checkpoint_name` calls, this code
arbitrarily uses the latter name (I think). Moreover there are probably ways
the name doesn't get picked up even if it should. A more thorough version might
check all uses, even ones that occur outside a higher-order primitive or
something. But for now this change makes the output slightly nicer for some
useful cases.