Remove uses of jax.experimental.host_callback.call
The jax.experimental.host_callback module is deprecated and will be removed.

See https://github.com/google/jax/issues/20385.

Most of the changes here have to do with the fact that io_callback does not pass the `device` to the callback. Fortunately, it seems that this code uses the device argument only for logging. I removed all uses of `device`.

PiperOrigin-RevId: 618402363