Transpose for reduce window.
Implements the transpose rule for general reduce window support. We do this by padding out the cotangents and running a reduce_window on the cotangents, using mostly the same windowing properties, with the specific difference in how we approach dilation and padding during the reduction (as these are accounted for during the padding of the cotangents).

Note - that this CL highlights a small bug/issue in the tracing mechanism itself. We cannot reuse the primals side of the JVP, and are forced to recompute it, due to the fact that attempting to use primals from the initial JVP computation flags the primals side arrays as unknown.

PiperOrigin-RevId: 632679634