Cleanup token handling during lowering
Version 0.4.27 of jaxlib is now the minimum version and it supports
real stablehlo tokens as module inputs and outputs. Hence we can
now clean up `mlir.lower_jaxpr_to_fun` to not use the kwargs
`create_tokens` and `replace_tokens_with_dummy` (both of them
are always False now).

We also remove `num_output_tokens` that is not used.