jax.nn.softmax: only use initial=-inf when necessary
The reason we need initial at all here is becuase when a user passes `where` or takes the softmax of an empty array, it is required. In other cases, the value does not affect the output, and may lead to unnecessary computation.

Partially addresses https://github.com/google/jax/issues/20856.

PiperOrigin-RevId: 634465693