Add RegularGridInterpolator to generated API docs
In responding to gh21279, I noticed that `RegularGridInterpolator` isn't
currently listed in the API docs. I know that `scipy.interpolate` is out
of scope (https://jax.readthedocs.io/en/latest/jep/18137-numpy-scipy-scope.html#scipy-interpolate),
but since we do currently provide this wrapper, it seems like it makes
sense to include it in the docs!