[Mosaic] Use implicit shape over directly using `implicit_dim()` in layout.{cc, h}
This CL also fixes a bug in `VectorLayout::join` where `ImplicitDim::kMinor` was considered equivalent to `ImplicitDim::kNone` when the shape's minor dimension is 1 (also needed to check that the second-minor dimension is 1).

Often handling every implicit dim case separately is more complex and error-prone.

There will be more follow-up changes to do this consistently elsewhere the code. This is also a first step towards 0D layout support.

PiperOrigin-RevId: 636250759