[Mosaic GPU] Implement basic support for aliasing shared memory.
When the user constructs the relevant shapes that live in `smem` throughout the
program, they now have the possibility of using a `mosaic_gpu.Union` of PyTrees
instead of a single `PyTree`.

`mosaic_gpu.Union` allows declaring several sets of buffers where within the
set, the buffers are alive concurrently, but between two distinct sets, the
buffers are alive for non-intersecting time intervals.

PiperOrigin-RevId: 636533045